backend/requirements.txt
aniso8601==9.0.1
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.1
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-camel-case==1.4.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
exceptiongroup==1.2.0
Faker==12.0.1
graphene==3.3
graphene-django==3.2.0
graphql-core==3.2.3
graphql-relay==3.2.0
gunicorn==21.2.0
idna==3.6
inflection==0.5.1
iniconfig==2.0.0
Markdown==3.5.2
mixer==7.2.2
packaging==23.2
pluggy==1.4.0
promise==2.3
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytest==8.0.0
pytest-django==4.8.0
pytest-mock==3.12.0
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
tomli==2.0.1
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.2.0
